pyproject.toml
[tool.poetry]
name = "cosbot"
version = "0.1.0"
description = ""
authors = ["Andrew <andrewbnx@outlook.com>"]
readme = "README.md"
packages = [{include = "bot"}]

[tool.poetry.dependencies]
python = "^3.12"
discord-py = {extras = ["speed", "voice"], version = "^2.4.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
pydantic = "^2.10.1"
redis = {extras = ["hiredis"], version = "^5.2.0"}
aiohttp = "^3.11.7"
pydantic-settings = "^2.6.1"
aiosignal = "^1.3.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
pylint = "^3.3.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
update-deps = "scripts.dependencies:update_dependencies"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "C0330",  # bad-continuation
    "C0326",  # bad-whitespace
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true